Move colorizes icons to their own file
