+++ packages/mobile/src/components/now-playing-drawer/AudioControls.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useMemo } from 'react'
+import { useCallback } from 'react'
 
 import { getRepeat, getShuffle } from 'common/store/queue/selectors'
 import { shuffle, repeat } from 'common/store/queue/slice'
@@ -6,18 +6,11 @@ import { RepeatMode } from 'common/store/queue/types'
 import { Animated, View, StyleSheet } from 'react-native'
 import { useDispatch, useSelector } from 'react-redux'
 
-import IconPause from 'app/assets/animations/iconPause.json'
-import IconPlay from 'app/assets/animations/iconPlay.json'
-import IconRepeatAllLight from 'app/assets/animations/iconRepeatAllLight.json'
-import IconRepeatOffLight from 'app/assets/animations/iconRepeatOffLight.json'
-import IconRepeatSingleLight from 'app/assets/animations/iconRepeatSingleLight.json'
-import IconShuffleOffLight from 'app/assets/animations/iconShuffleOffLight.json'
-import IconShuffleOnLight from 'app/assets/animations/iconShuffleOnLight.json'
 import IconNext from 'app/assets/images/iconNext.svg'
 import IconPodcastBack from 'app/assets/images/iconPodcastBack.svg'
 import IconPodcastForward from 'app/assets/images/iconPodcastForward.svg'
 import IconPrev from 'app/assets/images/iconPrev.svg'
-import { AnimatedButton, IconButton } from 'app/components/core'
+import { IconButton } from 'app/components/core'
 import * as haptics from 'app/haptics'
 import { useDispatchWeb } from 'app/hooks/useDispatchWeb'
 import { usePressScaleAnimation } from 'app/hooks/usePressScaleAnimation'
@@ -25,8 +18,11 @@ import { useSelectorWeb } from 'app/hooks/useSelectorWeb'
 import { useThemedStyles } from 'app/hooks/useThemedStyles'
 import { pause, play } from 'app/store/audio/actions'
 import { getPlaying } from 'app/store/audio/selectors'
-import { colorize } from 'app/utils/colorizeLottie'
-import { ThemeColors, useThemeColors } from 'app/utils/theme'
+import { ThemeColors } from 'app/utils/theme'
+
+import { PlayButton } from './PlayButton'
+import { RepeatButton } from './RepeatButton'
+import { ShuffleButton } from './ShuffleButton'
 
 const createStyles = (themeColors: ThemeColors) =>
   StyleSheet.create({
@@ -71,7 +67,6 @@ export const AudioControls = ({
   const dispatchWeb = useDispatchWeb()
 
   const styles = useThemedStyles(createStyles)
-  const { background, neutral, primary } = useThemeColors()
 
   const isPlaying = useSelector(getPlaying)
   const shuffleEnabled = useSelectorWeb(getShuffle)
@@ -118,197 +113,9 @@ export const AudioControls = ({
     dispatchWeb(repeat({ mode }))
   }, [dispatchWeb, repeatMode])
 
-  const ColorizedRepeatAllIcon = useMemo(
-    () =>
-      colorize(IconRepeatAllLight, {
-        // repeat number Outlines.Group 1.Fill 1
-        'assets.0.layers.0.shapes.0.it.1.c.k': background,
-        // repeat number Outlines.Group 2.Fill 1
-        'assets.0.layers.0.shapes.1.it.1.c.k': primary,
-        // repeat number Outlines.Group 3.Fill 1
-        'assets.0.layers.0.shapes.2.it.1.c.k': background,
-        // repeat number Outlines.Group 4.Fill 1
-        'assets.0.layers.0.shapes.3.it.1.c.k': '#000000',
-        // repeat Outlines.Group 2.Fill 1
-        'assets.0.layers.1.shapes.0.it.1.c.k.0.s': neutral,
-        // repeat Outlines.Group 2.Fill 1
-        'assets.0.layers.1.shapes.0.it.1.c.k.0.e': primary,
-        // repeat Outlines.Group 2.Fill 1
-        'assets.0.layers.1.shapes.0.it.1.c.k.1.s': primary,
-        // repeat Outlines.Group 2.Fill 1
-        'assets.0.layers.1.shapes.0.it.1.c.k.1.e': primary,
-        // repeat Outlines.Group 2.Fill 1
-        'assets.0.layers.1.shapes.0.it.1.c.k.2.s': primary,
-        // repeat Outlines.Group 2.Fill 1
-        'assets.0.layers.1.shapes.0.it.1.c.k.2.e': neutral,
-        // repeat number Outlines.Group 1.Fill 1
-        'layers.1.shapes.0.it.1.c.k': background,
-        // repeat number Outlines.Group 2.Fill 1
-        'layers.1.shapes.1.it.1.c.k': primary,
-        // repeat number Outlines.Group 3.Fill 1
-        'layers.1.shapes.2.it.1.c.k': background,
-        // repeat number Outlines.Group 4.Fill 1
-        'layers.1.shapes.3.it.1.c.k': '#000000',
-        // repeat Outlines.Group 2.Fill 1
-        'layers.2.shapes.0.it.1.c.k.0.s': neutral,
-        // repeat Outlines.Group 2.Fill 1
-        'layers.2.shapes.0.it.1.c.k.0.e': primary,
-        // repeat Outlines.Group 2.Fill 1
-        'layers.2.shapes.0.it.1.c.k.1.s': primary,
-        // repeat Outlines.Group 2.Fill 1
-        'layers.2.shapes.0.it.1.c.k.1.e': primary,
-        // repeat Outlines.Group 2.Fill 1
-        'layers.2.shapes.0.it.1.c.k.2.s': primary,
-        // repeat Outlines.Group 2.Fill 1
-        'layers.2.shapes.0.it.1.c.k.2.e': neutral
-      }),
-    [background, neutral, primary]
-  )
-
-  const ColorizedRepeatSingleIcon = useMemo(
-    () =>
-      colorize(IconRepeatSingleLight, {
-        // repeat number Outlines.Group 1.Fill 1
-        'layers.0.shapes.0.it.1.c.k': background,
-        // repeat number Outlines.Group 2.Fill 1
-        'layers.0.shapes.1.it.1.c.k': primary,
-        // repeat number Outlines.Group 3.Fill 1
-        'layers.0.shapes.2.it.1.c.k': background,
-        // repeat number Outlines.Group 4.Fill 1
-        'layers.0.shapes.3.it.1.c.k': '#000000',
-        // repeat Outlines.Group 2.Fill 1
-        'layers.1.shapes.0.it.1.c.k.0.s': neutral,
-        // repeat Outlines.Group 2.Fill 1
-        'layers.1.shapes.0.it.1.c.k.0.e': primary,
-        // repeat Outlines.Group 2.Fill 1
-        'layers.1.shapes.0.it.1.c.k.1.s': primary,
-        // repeat Outlines.Group 2.Fill 1
-        'layers.1.shapes.0.it.1.c.k.1.e': primary,
-        // repeat Outlines.Group 2.Fill 1
-        'layers.1.shapes.0.it.1.c.k.2.s': primary,
-        // repeat Outlines.Group 2.Fill 1
-        'layers.1.shapes.0.it.1.c.k.2.e': neutral
-      }),
-    [background, neutral, primary]
-  )
-
-  const ColorizedRepeatOffIcon = useMemo(
-    () =>
-      colorize(IconRepeatOffLight, {
-        // repeat number Outlines.Group 1.Fill 1
-        'layers.0.shapes.0.it.1.c.k': background,
-        // repeat number Outlines.Group 2.Fill 1
-        'layers.0.shapes.1.it.1.c.k': primary,
-        // repeat number Outlines.Group 3.Fill 1
-        'layers.0.shapes.2.it.1.c.k': background,
-        // repeat number Outlines.Group 4.Fill 1
-        'layers.0.shapes.3.it.1.c.k': '#000000',
-        // repeat Outlines.Group 2.Fill 1
-        'layers.1.shapes.0.it.1.c.k.0.s': neutral,
-        // repeat Outlines.Group 2.Fill 1
-        'layers.1.shapes.0.it.1.c.k.0.e': primary,
-        // repeat Outlines.Group 2.Fill 1
-        'layers.1.shapes.0.it.1.c.k.1.s': primary,
-        // repeat Outlines.Group 2.Fill 1
-        'layers.1.shapes.0.it.1.c.k.1.e': primary,
-        // repeat Outlines.Group 2.Fill 1
-        'layers.1.shapes.0.it.1.c.k.2.s': primary,
-        // repeat Outlines.Group 2.Fill 1
-        'layers.1.shapes.0.it.1.c.k.2.e': neutral
-      }),
-    [background, neutral, primary]
-  )
-
-  const repeatIconJSON = [
-    ColorizedRepeatAllIcon,
-    ColorizedRepeatSingleIcon,
-    ColorizedRepeatOffIcon
-  ]
-
-  const ColorizedShuffleOnIcon = useMemo(
-    () =>
-      colorize(IconShuffleOnLight, {
-        // Arrow 2.Head.Fill 1
-        'layers.0.shapes.0.it.2.c.k.0.s': neutral,
-        // Arrow 2.Head.Fill 1
-        'layers.0.shapes.0.it.2.c.k.0.e': primary,
-        // Arrow 2.Tail.Stroke 1
-        'layers.0.shapes.1.it.1.c.k.0.s': neutral,
-        // Arrow 2.Tail.Stroke 1
-        'layers.0.shapes.1.it.1.c.k.0.e': primary,
-        // Arrow 1.Head.Fill 1
-        'layers.1.shapes.0.it.2.c.k.0.s': neutral,
-        // Arrow 1.Head.Fill 1
-        'layers.1.shapes.0.it.2.c.k.0.e': primary,
-        // Arrow 1.Tail.Stroke 1
-        'layers.1.shapes.1.it.1.c.k.0.s': neutral,
-        // Arrow 1.Tail.Stroke 1
-        'layers.1.shapes.1.it.1.c.k.0.e': primary
-      }),
-    [neutral, primary]
-  )
-
-  const ColorizedShuffleOffIcon = useMemo(
-    () =>
-      colorize(IconShuffleOffLight, {
-        // Arrow 2.Head.Fill 1
-        'layers.0.shapes.0.it.2.c.k.0.s': primary,
-        // Arrow 2.Head.Fill 1
-        'layers.0.shapes.0.it.2.c.k.0.e': neutral,
-        // Arrow 2.Tail.Stroke 1
-        'layers.0.shapes.1.it.1.c.k.0.s': primary,
-        // Arrow 2.Tail.Stroke 1
-        'layers.0.shapes.1.it.1.c.k.0.e': neutral,
-        // Arrow 1.Head.Fill 1
-        'layers.1.shapes.0.it.2.c.k.0.s': primary,
-        // Arrow 1.Head.Fill 1
-        'layers.1.shapes.0.it.2.c.k.0.e': neutral,
-        // Arrow 1.Tail.Stroke 1
-        'layers.1.shapes.1.it.1.c.k.0.s': primary,
-        // Arrow 1.Tail.Stroke 1
-        'layers.1.shapes.1.it.1.c.k.0.e': neutral
-      }),
-    [neutral, primary]
-  )
-
-  const shuffleIconJSON = [ColorizedShuffleOnIcon, ColorizedShuffleOffIcon]
-
-  const ColorizedPlayIcon = useMemo(
-    () =>
-      colorize(IconPlay, {
-        // #playpause1.Group 1.Fill 1
-        'layers.0.shapes.0.it.1.c.k': background,
-        // #playpause2.Left.Fill 1
-        'layers.1.shapes.0.it.1.c.k': background,
-        // #playpause2.Right.Fill 1
-        'layers.1.shapes.1.it.1.c.k': background,
-        // #primaryBG.Group 2.Fill 1
-        'layers.2.shapes.0.it.1.c.k': primary
-      }),
-    [background, primary]
-  )
-
-  const ColorizedPauseIcon = useMemo(
-    () =>
-      colorize(IconPause, {
-        // #playpause1.Group 1.Fill 1
-        'layers.0.shapes.0.it.1.c.k': background,
-        // #playpause2.Left.Fill 1
-        'layers.1.shapes.0.it.1.c.k': background,
-        // #playpause2.Right.Fill 1
-        'layers.1.shapes.1.it.1.c.k': background,
-        // #primaryBG.Group 2.Fill 1
-        'layers.2.shapes.0.it.1.c.k': primary
-      }),
-    [background, primary]
-  )
-
-  const playButtonIconJSON = [ColorizedPlayIcon, ColorizedPauseIcon]
-
   const renderRepeatButton = () => {
     return (
-      <AnimatedButton
-        iconJSON={repeatIconJSON}
+      <RepeatButton
         onPress={onPressRepeat}
         style={styles.button}
         wrapperStyle={styles.shuffleRepeatIcons}
@@ -327,9 +134,8 @@ export const AudioControls = ({
   const renderPlayButton = () => {
     return (
       <Animated.View style={{ transform: [{ scale }] }}>
-        <AnimatedButton
+        <PlayButton
           iconIndex={isPlaying ? 1 : 0}
-          iconJSON={playButtonIconJSON}
           onPress={onPressPlayButton}
           onPressIn={handlePressInScale}
           onPressOut={handlePressOutScale}
@@ -350,8 +156,7 @@ export const AudioControls = ({
   }
   const renderShuffleButton = () => {
     return (
-      <AnimatedButton
-        iconJSON={shuffleIconJSON}
+      <ShuffleButton
         onPress={onPressShuffle}
         style={styles.button}
         wrapperStyle={styles.shuffleRepeatIcons}
